+++ src/webeffect/index.js
@@ -1,0 +1,28 @@
+export {
+  CanonicalEventRegistry,
+  STANDARD_CANONICAL_EVENTS,
+} from "./eventRegistry.js";
+
+export {
+  detectEventDuplicates,
+} from "./duplicateDetection.js";
+
+export {
+  validateAttributionContinuity,
+} from "./attributionContinuity.js";
+
+export {
+  observeConsentStateDifferential,
+} from "./consentDifferential.js";
+
+export {
+  validateCommerceConsistency,
+} from "./commerceConsistency.js";
+
+export {
+  buildThirdPartyDependencyGraph,
+} from "./dependencyGraph.js";
+
+export {
+  webeffectDomainModule,
+} from "./domainModule.js";